refactor: extract agents, memory/edges, and findings/stats MCP tools into lithos.tools
First extraction slice (10 of 37 tools) out of LithosServer._register_tools,
proving the pattern on the domains that exercise late-bound components
(server.memory/server.projection) and the stats gauge-cache mutation:

- new Entrypoints-tier package lithos/tools/ with per-domain register(mcp,
  server) modules (agents, memory_edges, findings_stats) and register_all;
  handlers close over the server only and read components at call time —
  the late-binding rule is documented on the package
- new tools/_seam.py tool_span decorator owns the lithos.tool.<short> span
  convention and the CoordinationError -> canonical-envelope mapping; sits
  below @tool_metrics so mapped errors are not counted as tool errors
  (same semantics as the per-handler code it replaces)
- new telemetry.get_current_span() accessor (no-op span when OTEL inactive)
  so handler bodies never import opentelemetry directly
- lithos_conflict_resolve and lithos_cache_lookup gain spans they never had;
  early-invalid returns now happen inside the span (log-trace correlation) —
  accepted observability deltas, wire output unchanged
- test_all_server_tools_have_tool_metrics_decorator becomes a dual scan
  (remaining _register_tools source + every lithos/tools/ module)
- architecture.toml maps lithos.tools into Entrypoints; diagrams regenerated
- new tests/test_tool_seam.py covers the decorator stack contract (wraps/
  signature preservation, mapped-vs-raised metrics semantics)

Envelope-conformance suite passes unmodified (move-purity guardrail).

Part of the 2026-07 architecture review (tracker task 9f2be54f, PR 3 of 6).
