+++ src/lithos/server.py
@@ -10,7 +10,6 @@ import logging
 import math
 import time
 from datetime import UTC, datetime, timedelta
-from pathlib import Path
 from typing import Any
 
 from fastmcp import FastMCP
@@ -29,8 +28,6 @@ from lithos.envelopes import (
 )
 from lithos.errors import CoordinationError, SearchBackendError
 from lithos.events import (
-    AGENT_REGISTERED,
-    FINDING_POSTED,
     TASK_CANCELLED,
     TASK_CLAIMED,
     TASK_COMPLETED,
@@ -66,6 +63,7 @@ from lithos.telemetry import (
     register_sse_active_clients_observer,
     tool_metrics,
 )
+from lithos.tools import register_all
 from lithos.watch_intake import WatchIntake
 
 logger = logging.getLogger(__name__)
@@ -183,7 +181,9 @@ class LithosServer:
             self._config.events.max_sse_clients
         )
 
-        # Register all tools
+        # Register all tools: extracted domain modules first (lithos.tools),
+        # then the handlers not yet migrated out of _register_tools.
+        register_all(self.mcp, self)
         self._register_tools()
 
         # Mount SSE delivery endpoint
@@ -1964,178 +1964,6 @@ class LithosServer:
 
         @self.mcp.tool()
         @tool_metrics()
-        async def lithos_edge_upsert(
-            from_id: str,
-            to_id: str,
-            type: str,
-            weight: float,
-            namespace: str,
-            provenance_actor: str | None = None,
-            provenance_type: str | None = None,
-            evidence: Any = None,
-            conflict_state: str | None = None,
-        ) -> dict[str, Any]:
-            """Create or update a typed edge in edges.db.
-
-            Upsert key is (from_id, to_id, type, namespace).
-
-            Args:
-                from_id: Source node ID
-                to_id: Target node ID
-                type: Edge type (e.g. 'derived_from', 'related_to')
-                weight: Edge weight (float)
-                namespace: Namespace for the edge (required)
-                provenance_actor: Agent/process that created the edge
-                provenance_type: How the edge was derived
-                evidence: Supporting evidence (dict or list only, not scalars)
-                conflict_state: Conflict state marker
-
-            Returns:
-                Status envelope with edge_id.
-            """
-            logger.info("lithos_edge_upsert from=%s to=%s type=%s", from_id, to_id, type)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.edge_upsert") as span:
-                span.set_attribute("lithos.tool", "lithos_edge_upsert")
-
-                if not namespace:
-                    return error_envelope("invalid_input", "namespace is required")
-
-                # Validate evidence type
-                if evidence is not None and not isinstance(evidence, (dict, list)):
-                    return error_envelope(
-                        "invalid_input",
-                        "evidence must be a dict, list, or null — scalars are not accepted",
-                    )
-
-                evidence_str = json.dumps(evidence) if evidence is not None else None
-
-                # The MCP surface carries no separate ``agent`` field for
-                # this tool; the asserting party is ``provenance_actor`` (with
-                # a tool-name fallback so ``ensure_agent_known`` always runs,
-                # per ADR-0006 Slice 1 / issue #263).
-                agent = provenance_actor or "lithos_edge_upsert"
-                edge_id = await self.memory.edge_upsert(
-                    agent=agent,
-                    from_id=from_id,
-                    to_id=to_id,
-                    edge_type=type,
-                    weight=weight,
-                    namespace=namespace,
-                    provenance_actor=provenance_actor,
-                    provenance_type=provenance_type,
-                    evidence=evidence_str,
-                    conflict_state=conflict_state,
-                )
-
-                return {
-                    "status": "ok",
-                    "edge_id": edge_id,
-                }
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_edge_list(
-            from_id: str | None = None,
-            to_id: str | None = None,
-            type: str | None = None,
-            namespace: str | None = None,
-        ) -> dict[str, Any]:
-            """Query edges from edges.db by optional filters.
-
-            Args:
-                from_id: Filter by source node ID
-                to_id: Filter by target node ID
-                type: Filter by edge type
-                namespace: Filter by namespace
-
-            Returns:
-                Dict with results list of edge dicts.
-            """
-            logger.info(
-                "lithos_edge_list from=%s to=%s type=%s ns=%s", from_id, to_id, type, namespace
-            )
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.edge_list") as span:
-                span.set_attribute("lithos.tool", "lithos_edge_list")
-
-                edges = await self.memory.edge_list(
-                    from_id=from_id,
-                    to_id=to_id,
-                    edge_type=type,
-                    namespace=namespace,
-                )
-
-                span.set_attribute("lithos.result_count", len(edges))
-                return {"results": edges}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_conflict_resolve(
-            edge_id: str,
-            resolution: str,
-            resolver: str,
-            winner_id: str | None = None,
-        ) -> dict[str, Any]:
-            """Resolve a contradiction between two notes.
-
-            Sets the resolution state on a contradicts edge so future retrieval
-            reflects the resolution.
-
-            Args:
-                edge_id: The edge ID of the contradiction to resolve
-                resolution: One of: accepted_dual, superseded, refuted, merged
-                resolver: Agent or user identifier performing the resolution
-                winner_id: Required when resolution is 'superseded'; must be
-                    either from_id or to_id of the edge
-            """
-            return await self.memory.conflict_resolve(
-                edge_id=edge_id,
-                resolution=resolution,
-                resolver=resolver,
-                winner_id=winner_id,
-            )
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_cache_lookup(
-            query: str,
-            source_url: str | None = None,
-            max_age_hours: float | None = None,
-            min_confidence: float = 0.5,
-            limit: int = 3,
-            tags: list[str] | None = None,
-        ) -> dict[str, Any]:
-            """Check if fresh cached knowledge exists before doing expensive research.
-
-            Returns a cache hit with full document content if fresh knowledge exists,
-            a stale reference if expired knowledge exists (update instead of duplicate),
-            or a clean miss if nothing relevant is found.
-
-            Args:
-                query: What you are about to research
-                source_url: Canonical URL for exact dedup-aware lookup
-                max_age_hours: Reject docs older than N hours (uses updated_at)
-                min_confidence: Minimum confidence score threshold — candidates whose
-                    ``metadata.confidence`` is strictly below this value are skipped
-                    entirely (default: 0.5).
-                limit: Max candidate docs to evaluate (default: 3).
-                tags: Restrict to tagged docs (AND semantics)
-
-            Returns:
-                Dict with hit, document, stale_exists, stale_id
-            """
-            return await self.memory.cache_lookup(
-                query=query,
-                source_url=source_url,
-                max_age_hours=max_age_hours,
-                min_confidence=min_confidence,
-                limit=limit,
-                tags=tags,
-            )
-
-        @self.mcp.tool()
-        @tool_metrics()
         async def lithos_list(
             path_prefix: str | None = None,
             tags: list[str] | None = None,
@@ -2494,135 +2322,6 @@ class LithosServer:
                     tags = {k: v for k, v in tags.items() if k.lower().startswith(prefix.lower())}
                 span.set_attribute("lithos.tag_count", len(tags))
                 return {"tags": tags}
-
-        # ==================== Agent Tools ====================
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_agent_register(
-            id: str,
-            name: str | None = None,
-            type: str | None = None,
-            metadata: dict[str, Any] | None = None,
-        ) -> dict[str, bool]:
-            """Register or update an agent.
-
-            Args:
-                id: Agent identifier
-                name: Human-friendly display name
-                type: Agent type (e.g., "agent-zero", "claude-code")
-                metadata: Optional extra info
-
-            Returns:
-                Dict with success and created booleans
-            """
-            logger.info("lithos_agent_register id=%s type=%s", id, type)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.agent_register") as span:
-                span.set_attribute("lithos.tool", "lithos_agent_register")
-                span.set_attribute("lithos.agent.id", id)
-                success, created = await self.coordination.register_agent(
-                    agent_id=id,
-                    name=name,
-                    agent_type=type,
-                    metadata=metadata,
-                )
-                span.set_attribute("lithos.created", created)
-
-                if success:
-                    await self._emit(
-                        LithosEvent(
-                            type=AGENT_REGISTERED,
-                            agent=id,
-                            payload={"agent_id": id, "name": name or ""},
-                        )
-                    )
-
-                return {"success": success, "created": created}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_agent_info(
-            id: str,
-        ) -> dict[str, Any] | None:
-            """Get agent information.
-
-            Args:
-                id: Agent identifier
-
-            Returns:
-                Agent info dict or None if not found
-            """
-            logger.info("lithos_agent_info id=%s", id)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.agent_info") as span:
-                span.set_attribute("lithos.tool", "lithos_agent_info")
-                span.set_attribute("lithos.agent.id", id)
-                agent = await self.coordination.get_agent(id)
-                if not agent:
-                    return None
-
-                return {
-                    "id": agent.id,
-                    "name": agent.name,
-                    "type": agent.type,
-                    "first_seen_at": (
-                        agent.first_seen_at.isoformat() if agent.first_seen_at else None
-                    ),
-                    "last_seen_at": (
-                        agent.last_seen_at.isoformat() if agent.last_seen_at else None
-                    ),
-                    "metadata": agent.metadata,
-                }
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_agent_list(
-            type: str | None = None,
-            active_since: str | None = None,
-        ) -> dict[str, list[dict[str, Any]]]:
-            """List all known agents.
-
-            Args:
-                type: Filter by agent type
-                active_since: Filter by last activity (ISO datetime)
-
-            Returns:
-                Dict with agents list
-            """
-            logger.info("lithos_agent_list type=%s", type)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.agent_list") as span:
-                span.set_attribute("lithos.tool", "lithos_agent_list")
-
-                since_dt = None
-                if active_since:
-                    try:
-                        since_dt = _normalize_datetime(datetime.fromisoformat(active_since))
-                    except ValueError:
-                        return invalid_input_envelope(
-                            f"Invalid active_since datetime: {active_since}"
-                        )
-
-                agents = await self.coordination.list_agents(
-                    agent_type=type,
-                    active_since=since_dt,
-                )
-
-                span.set_attribute("lithos.result_count", len(agents))
-                return {
-                    "agents": [
-                        {
-                            "id": a.id,
-                            "name": a.name,
-                            "type": a.type,
-                            "last_seen_at": (
-                                a.last_seen_at.isoformat() if a.last_seen_at else None
-                            ),
-                        }
-                        for a in agents
-                    ]
-                }
 
         # ==================== Coordination Tools ====================
 
@@ -3617,204 +3316,6 @@ class LithosServer:
                     return error_envelope("task_not_found", f"Task '{task_id}' not found.")
 
                 return {"task": _serialize_task_record(task)}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_finding_post(
-            task_id: str,
-            agent: str,
-            summary: str,
-            knowledge_id: str | None = None,
-        ) -> dict[str, str]:
-            """Post a finding to a task.
-
-            Args:
-                task_id: Task ID
-                agent: Agent posting the finding
-                summary: Finding summary
-                knowledge_id: Optional linked knowledge document ID
-
-            Returns:
-                Dict with finding_id
-            """
-            logger.info("lithos_finding_post task=%s agent=%s", task_id, agent)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.finding_post") as span:
-                span.set_attribute("lithos.tool", "lithos_finding_post")
-                span.set_attribute("lithos.agent", agent)
-                span.set_attribute("lithos.task_id", task_id)
-                finding_id = await self.coordination.post_finding(
-                    task_id=task_id,
-                    agent=agent,
-                    summary=summary,
-                    knowledge_id=knowledge_id,
-                )
-                span.set_attribute("lithos.finding_id", finding_id)
-
-                payload: dict[str, str | int | float | bool | None] = {
-                    "finding_id": finding_id,
-                    "task_id": task_id,
-                    "agent": agent,
-                }
-                if knowledge_id is not None:
-                    payload["knowledge_id"] = knowledge_id
-
-                await self._emit(
-                    LithosEvent(
-                        type=FINDING_POSTED,
-                        agent=agent,
-                        payload=payload,
-                    )
-                )
-
-                return {"finding_id": finding_id}
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_finding_list(
-            task_id: str,
-            since: str | None = None,
-        ) -> dict[str, list[dict[str, Any]]]:
-            """List findings for a task.
-
-            Args:
-                task_id: Task ID
-                since: Filter by created since (ISO datetime)
-
-            Returns:
-                Dict with findings list
-            """
-            logger.info("lithos_finding_list task=%s", task_id)
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.finding_list") as span:
-                span.set_attribute("lithos.tool", "lithos_finding_list")
-                span.set_attribute("lithos.task_id", task_id)
-
-                since_dt = None
-                if since:
-                    try:
-                        since_dt = _normalize_datetime(datetime.fromisoformat(since))
-                    except ValueError:
-                        return invalid_input_envelope(f"Invalid since datetime: {since}")
-
-                findings = await self.coordination.list_findings(
-                    task_id=task_id,
-                    since=since_dt,
-                )
-
-                span.set_attribute("lithos.result_count", len(findings))
-                return {
-                    "findings": [
-                        {
-                            "id": f.id,
-                            "agent": f.agent,
-                            "summary": f.summary,
-                            "knowledge_id": f.knowledge_id,
-                            "created_at": (f.created_at.isoformat() if f.created_at else None),
-                        }
-                        for f in findings
-                    ]
-                }
-
-        # ==================== System Tools ====================
-
-        @self.mcp.tool()
-        @tool_metrics()
-        async def lithos_stats() -> dict[str, Any]:
-            """Get knowledge base statistics and health indicators.
-
-            Returns:
-                Dict with document counts, search index stats, coordination
-                stats, and health signals (index drift, broken links, expired
-                claims, last-updated timestamps).
-            """
-            logger.info("lithos_stats")
-            tracer = get_tracer()
-            with tracer.start_as_current_span("lithos.tool.stats") as span:
-                span.set_attribute("lithos.tool", "lithos_stats")
-
-                # Get document count (from in-memory cache — always available)
-                total_docs = self.knowledge.document_count
-
-                # Semantic chunk count via the public surface (returns 0 when
-                # the Chroma store is quarantined).
-                chroma_chunk_count: int = self.search.count_chunks()
-
-                # Full-text document count with None-sentinel on failure.
-                # NOTE: this deliberately diverges from _safe_tantivy_count(),
-                # which returns *0* on error (used by OTEL gauge probes where a
-                # numeric value is always required).  Here we return *None* so
-                # callers can distinguish "index not yet written / unavailable"
-                # from "zero documents indexed" — a meaningful difference for
-                # drift detection and health reporting.  If you need the 0-on-
-                # error behaviour, use _safe_tantivy_count() instead.
-                tantivy_doc_count: int | None
-                try:
-                    tantivy_doc_count = self.search.count_documents()
-                except Exception:
-                    tantivy_doc_count = None
-
-                # Index drift: knowledge corpus vs Tantivy index
-                index_drift_detected = (
-                    tantivy_doc_count is not None and tantivy_doc_count != total_docs
-                )
-
-                # Get coordination stats
-                coord_stats = await self.coordination.get_stats()
-
-                # Update cached fields for synchronous OTEL gauge callbacks
-                self._cached_active_claims = coord_stats.get("open_claims", 0)
-                self._cached_agent_count = coord_stats.get("agents", 0)
-
-                # Get tag count
-                tags = await self.knowledge.get_all_tags()
-
-                # Unresolved wiki-links: nodes in the graph that have no
-                # matching document (represented as __unresolved__ placeholders)
-                graph_stats = self.graph.get_stats()
-                unresolved_links: int = graph_stats.get("unresolved_links", 0)
-
-                # Stale (expired) documents
-                expired_docs = self.knowledge.stale_document_count
-
-                # Index last-updated timestamps from filesystem mtime
-                def _dir_mtime(path: Path) -> str | None:
-                    try:
-                        mtime = path.stat().st_mtime
-                        return datetime.fromtimestamp(mtime, tz=UTC).isoformat()
-                    except OSError:
-                        return None
-
-                tantivy_last_updated = _dir_mtime(self._config.storage.tantivy_path)
-                chroma_last_updated = _dir_mtime(self._config.storage.chroma_path)
-
-                return {
-                    # Core counts
-                    "documents": total_docs,
-                    "chroma_chunk_count": chroma_chunk_count,
-                    "agents": coord_stats.get("agents", 0),
-                    "active_tasks": coord_stats.get("active_tasks", 0),
-                    "open_claims": coord_stats.get("open_claims", 0),
-                    "tags": len(tags),
-                    "duplicate_urls": self.knowledge.duplicate_url_count,
-                    # Health indicators
-                    "index_drift_detected": index_drift_detected,
-                    "tantivy_doc_count": tantivy_doc_count,
-                    "unresolved_links": unresolved_links,
-                    "expired_docs": expired_docs,
-                    "expired_claims": coord_stats.get("expired_claims", 0),
-                    "tantivy_last_updated": tantivy_last_updated,
-                    "chroma_last_updated": chroma_last_updated,
-                    # Graph stats
-                    "graph_node_count": graph_stats.get("nodes", 0),
-                    # graph_edge_count reflects ALL edges in the NetworkX graph,
-                    # including edges that point to __unresolved__ placeholder
-                    # nodes (i.e. wiki-links whose target document does not yet
-                    # exist).  It therefore equals resolved_edges + unresolved_links,
-                    # not just resolved edges.  Compare with unresolved_links above
-                    # to infer the resolved-only edge count if needed.
-                    "graph_edge_count": graph_stats.get("edges", 0),
-                }
 
 
 def _format_sse(event: LithosEvent) -> str:
